Move message list code to 'messagelist' package
