--- app/ui/legacy/src/main/java/com/fsck/k9/fragment/KoinModule.kt
@@ -1,7 +1,0 @@
-package com.fsck.k9.fragment
-
-import org.koin.dsl.module
-
-val fragmentModule = module {
-    single { SortTypeToastProvider() }
-}
