+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/messagelist/KoinModule.kt
@@ -17,4 +17,5 @@ val messageListUiModule = module {
     factory {
         MessageListLiveDataFactory(messageListLoader = get(), preferences = get(), messageListRepository = get())
     }
+    single { SortTypeToastProvider() }
 }
